refactor(Movie): changed 'popularity' from 'Float' to 'Double'.

+++ MobileChallenge01/Screens/Shared/Model/Movie.swift
@@ -12,6 +12,6 @@ struct Movie {
     let title: String
     let overview: String
     let releaseDate: String
-    let popularity: Float
+    let popularity: Double
     let genres: [String]
 }
